[rename] Modify a directory structure: /component/convert - seperate /component/character, /component/novel Github issue #34
